[package] ez-ipudate: fix config generator to permit default server
This patch allows ez-ipupdate to use the default server for the specified DDNS
service type when no server hostname is supplied by the user.

In detail:

The default config file supplied in the ez-ipupdate package does not include a
"server" line, and ez-ipupdate itself does not require an explicit server
(since it will select a server based on the service type).

Unfortunately, if the user simply fills in the blanks in the OpenWrt-supplied
ez-ipupdate config, the service will silently fail to start because the hotplug
 script inserts a line in the generated config which states "server=", and
ez-ipupdate doesn't like a blank server setting.

The above scenario doesn't seem like a good user experience, so this is a
small patch to the hotplug script to omit the "server" line (and let
ez-ipupdate select its own default) in the event that the user has not
specified a server.

Signed-off-by: Scott Dudley <spmsink@hotmail.com>