[buildroot] add support for darcs repositories
This patch adds support for darcs repositories (as sources of packages).

It does *not* add support for darcs:// URI scheme because such a scheme
do not exist (AFAIK).  You must therefore manually set PKG_SOURCE_PROTO
to darcs in your Makefile (and use a regular http:// URI). You also have
to set PKG_SOURCE_VERSION to a string matching a tag contained in the
repository, and PKG_SOURCE_SUBDIR to something sensible
($(PACKAGE_NAME)-$(PACKAGE_SOURCE_VERSION) for instance).

Same rationale as for the previous patch ("useless to most, but small
and straightforward so why no add it?).

Best regards,

Signed-off-by: Gabriel Kerneis <kerneis@pps.jussieu.fr>