[package] ddns-scripts: add https support to ddns-scripts
adds a "use_https" and "cacert" option to the ddns-scripts package.

when use_https is enabled, then curl is required to perform the update.

the cacert option can be set to either a single file containing a trust
signing authority certificate or a directory containing certificates
which has been prepared with the usual c_rehash method.

Signed-off-by: Matthew William Cox <matt@mattcox.ca>