[package] minidlna/ffmpeg: add minidlna profile to ffmpeg
This adds a profile to ffmpeg to support minidlna, similar to the
libdlna/ushare profile. When minidlna encounters media, it uses ffmpeg to
figure out what it is. If ffmpeg fails to open it, then minidlna will try
and fail to read the file on its own. The profile may need to be extended;
I attempted to cover all popular formats for dlna streaming.

Tested with the following container/codec combinations:

mkv with ac3/h264/srt
mp4 with aac/h264/ttxt
ogg vorbis
flac
mp3

Should also be able to handle, but untested:
mpeg
avi
xvid
jpeg
png
wmv1
wmv2

Signed-off-by: Ian Leonard <antonlacon_at_gmail.com>