mac80211: fix memory leak on bcm5354
The ucode triggers the gpio pins used for the buttons and causes diag to handle them and send them to userspace.
This only happened with a more recent firmware.
This patch was tested on an Asus wl-520GU and should also work on the other bcm5354 based asus devices. I do not know if it works on the dir320.