[lantiq] pci gpio mapping
This patch fixes the GPIO ALTSEL settings for some of the GPIOs used by the PCI subsystem in Lantiq Danube. These changes are required for more than one PCI device to work. Tested with an ARV7510PW having two PCI-devices; a VIA USB controller and a Ralink WLAN mini-PCI card.

Signed-off-by: Matti Laakso <malaakso@elisanet.fi>