ar71xx: add a helper function for external LNA GPIO setup
The built-in wireless MAC of the AR934x SoC can handle
external LNAs and the control signal of the LNAs can be
routed to any GPIO line. Add a helper function which
can be used to configure the GPIO lines.

The helper function will be used for AR934x boards
which are using externel LNAs to improve sensitivity.

Signed-off-by: Gabor Juhos <juhosg@openwrt.org>