x86_64: fix sysupgrade
Function move_config should be called after or during preinit_mount_root
hook in do_mount_root function. At this state sysupgrade.tgz is not in its
place during that time when do_mount_root is called. Function move_config
is called later so the sysupgrade.tgz stays in root directory to the second
restart when it is unpacked properly.

This patch adds move_config function to preinit_mount_root hook instead
of preinit_main and changes the filename of script to be called before
80_mount_root is called. It will prepare the sysupgrade archive for do_mount_root
within preinit_mount_root hook.

This issue was introduced for target x86. Since the target for x86_64 is
similar to x86, this issue is present also in this target. It solves the
ticket #15042 and #14088.

Signed-off-by: Jiri Slachta <slachta@cesnet.cz>
Tested-By: Stefan Hellermann <stefan@the2masters.de>