kobs-ng: Add support for overriding size
By default kobs-ng will use the size of /dev/mtd0 for the IMX bootstreams.
This patch allows overriding that size in order to reserve space at the end
of the mtd partition for u-boot.img in the case of using an SPL bootloader.

For example, to flash an SPL bootloader that loads u-boot.img from 0xe0000
(14MB) offset in a 16MB /dev/mtd0:

  kobs-ng init -v -x --search_exponent=1 --chip_0_size=0xe00000 SPL
  nandwrite --start=0xe00000 --pad /dev/mtd0 u-boot.img

Signed-off-by: Tim Harvey <tharvey@gateworks.com>