brcm47xx: do not use mac addresses with 00:90:4C prefix, prevent mac address collisions.
The address prefix 00:90:4C is used by Broadcom in their initial
configuration. When a mac address with the prefix 00:90:4C is used
all devices from the same series are sharing the same mac address.
To prevent mac address collisions we replace them with a mac address
based on the base address.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>