ar71xx: Support for Mikrotik RB2011UiAS
The RB2011UiAS is almost exactly the same as the RB2011UAS, which is
already supported. Hardware-wise, it adds a passive power injector on
Eth10. Software-wise, Routerboot passes a different board name on the
kernel command line. Hence the need for a patch.

Not yet working:

* touchscreen: doesn't work on RB2011UAS either. The touchscreen itself
  seems to be driven by an ADS7843-compatible IC, but figuring out which
  GPIO pins it's connected to is work for the future.
* control of power injector: remains off. Probably also GPIO-controlled.

Signed-off-by: Soren Harward <stharward@gmail.com>
Patchwork: http://patchwork.openwrt.org/patch/5841/
[juhosg: rename and refresh the kernel patch]
Signed-off-by: Gabor Juhos <juhosg@openwrt.org>