sdk: don't try to build in-kernel kmods
We don't ship the kernel sources, so using the base git as a feed will
fail when trying to build kernel modules with separate install steps.
Instead of trying to fixup the install steps, let's just skip building
kernel modules alltogether and just create empty packages.

Out-of-kernel modules are still expected to exist and are packaged, as
for these sources are fetched during the normal build steps.

Reported-by: Jo-Philipp Wich <jow@openwrt.org>
Signed-off-by: Jonas Gorski <jogo@openwrt.org>