bcm53xx: drop hacky patch making bcma init later
The only reason this patch was helping was moving bcma init *after*
of_serial init. It was a bit of luck because both drivers use
device_initcall.
It isn't really a correct way to provide early logging.
Now we have earlyprintk support in bcm53xx we can drop this patch.

Signed-off-by: Rafał Miłecki <zajec5@gmail.com>